pyproject.toml
[project]
name = "delivery-agua"
version = "0.1.0"
description = ""
authors = [
    {name = "Kauanrodrigues01m",email = "kauanrl09@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.9"
taskipy = "^1.14.1"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "B"]
ignore = []
exclude = ["migrations", ".venv", "venv", "env"]

[tool.taskipy.tasks]
run = "python manage.py runserver"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
createsuperuser = "python manage.py createsuperuser"
shell = "python manage.py shell"
test = "python manage.py test"
lint = 'ruff check .'
format = 'ruff check --fix && ruff format .'
